Housekeeping and adding more functions
