+++ src/fight.cpp
@@ -30,6 +30,7 @@
 #include "redit.hpp"
 #include "factions.hpp"
 #include "metrics.hpp"
+#include "gmcp.hpp"
 
 int initiative_until_global_reroll = 0;
 
@@ -270,6 +271,7 @@ bool update_pos(struct char_data * victim, bool protect_spells_from_purge)
       if (GET_BIOWARE_TYPE(bio) == BIO_PAINEDITOR && GET_BIOWARE_IS_ACTIVATED(bio)) {
         // Ensure they're not trapped in mort status by an active pain editor.
         GET_POS(victim) = MAX(POS_LYING, GET_POS(victim));
+        SendGMCPCharStatus(victim); 
         return FALSE;
       }
     }
@@ -281,6 +283,7 @@ bool update_pos(struct char_data * victim, bool protect_spells_from_purge)
 
   // Are they doing fine?
   else if ((GET_PHYSICAL(victim) >= 100) && (GET_POS(victim) > POS_STUNNED)) {
+    SendGMCPCharStatus(victim); 
     return FALSE;
   }
 
@@ -291,6 +294,7 @@ bool update_pos(struct char_data * victim, bool protect_spells_from_purge)
     }
 
     GET_POS(victim) = victim->in_veh ? POS_SITTING : POS_STANDING;
+    SendGMCPCharStatus(victim); 
     return FALSE;
   }
 
@@ -319,6 +323,7 @@ bool update_pos(struct char_data * victim, bool protect_spells_from_purge)
   GET_INIT_ROLL(victim) = 0;
 
   if (restore_to_full_health_if_still_in_chargen(victim)) {
+    SendGMCPCharStatus(victim); 
     return FALSE;
   }
 
@@ -353,6 +358,7 @@ bool update_pos(struct char_data * victim, bool protect_spells_from_purge)
     }
   }
 
+  SendGMCPCharStatus(victim); 
   return FALSE;
 }
 #undef MAKE_MORTALLY_WOUNDED
