+++ src/protocol.hpp
@@ -544,14 +544,6 @@ void MXPSendTag( descriptor_t *apDescriptor, const char *apTag );
 void SoundSend( descriptor_t *apDescriptor, const char *apTrigger );
 
 /******************************************************************************
- GMCP functions.
- ******************************************************************************/
-void SendGMCP( descriptor_t *apDescriptor, const char *module, const char *apData );
-void SendGMCPRoomInfo( struct char_data *ch, struct room_data *room );
-void SendGMCPCoreSupports ( descriptor_t *apDescriptor );
-void SendGMCPCharStatus( struct char_data * ch );
-
-/******************************************************************************
  Colour functions.
  ******************************************************************************/
